packages/outpostapi/requirements.txt
asgiref==3.6.0
Django==4.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
dnspython==2.3.0
psycopg2==2.9.6
pykerberos==1.2.4
python-snappy==0.6.1
pytz==2023.3
sqlparse==0.4.3
